+++ packages/jest-cli/src/__tests__/watch.test.js
@@ -86,6 +86,11 @@ jest.doMock(
   {virtual: true},
 );
 
+const regularUpdateGlobalConfig = require('../lib/update_global_config')
+  .default;
+const updateGlobalConfig = jest.fn(regularUpdateGlobalConfig);
+jest.doMock('../lib/update_global_config', () => updateGlobalConfig);
+
 const watch = require('../watch').default;
 
 const nextTick = () => new Promise(res => process.nextTick(res));
@@ -176,7 +181,7 @@ describe('Watch mode flows', () => {
       await watch(
         Object.assign({}, globalConfig, {
           rootDir: __dirname,
-          watchPlugins: [watchPluginPath],
+          watchPlugins: [{config: {}, path: watchPluginPath}],
         }),
         contexts,
         pipe,
@@ -195,7 +200,10 @@ describe('Watch mode flows', () => {
     ci_watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [watchPlugin2Path, watchPluginPath],
+        watchPlugins: [
+          {config: {}, path: watchPluginPath},
+          {config: {}, path: watchPlugin2Path},
+        ],
       }),
       contexts,
       pipe,
@@ -302,7 +310,7 @@ describe('Watch mode flows', () => {
     watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [pluginPath],
+        watchPlugins: [{config: {}, path: pluginPath}],
       }),
       contexts,
       pipe,
@@ -338,7 +346,7 @@ describe('Watch mode flows', () => {
     watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [pluginPath],
+        watchPlugins: [{config: {}, path: pluginPath}],
       }),
       contexts,
       pipe,
@@ -354,6 +362,47 @@ describe('Watch mode flows', () => {
     await nextTick();
 
     expect(run).toHaveBeenCalled();
+  });
+
+  it('allows WatchPlugins to be configured', async () => {
+    const pluginPath = `${__dirname}/__fixtures__/plugin_path_with_config`;
+    jest.doMock(
+      pluginPath,
+      () =>
+        class WatchPlugin {
+          constructor({config}) {
+            this._key = config.key;
+            this._prompt = config.prompt;
+          }
+          onKey() {}
+          run() {}
+          getUsageInfo() {
+            return {
+              key: this._key || 'z',
+              prompt: this._prompt || 'default prompt',
+            };
+          }
+        },
+      {virtual: true},
+    );
+
+    watch(
+      Object.assign({}, globalConfig, {
+        rootDir: __dirname,
+        watchPlugins: [
+          {
+            config: {key: 'k', prompt: 'filter with a custom prompt'},
+            path: pluginPath,
+          },
+        ],
+      }),
+      contexts,
+      pipe,
+      hasteMapInstances,
+      stdin,
+    );
+
+    expect(pipe.write.mock.calls.reverse()[0]).toMatchSnapshot();
   });
 
   it('allows WatchPlugins to hook into file system changes', async () => {
@@ -373,7 +422,7 @@ describe('Watch mode flows', () => {
     watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [pluginPath],
+        watchPlugins: [{config: {}, path: pluginPath}],
       }),
       contexts,
       pipe,
@@ -390,6 +439,101 @@ describe('Watch mode flows', () => {
       ],
     });
   });
+
+  it.each`
+    ok | option
+    ✔︎ | bail
+    ✖︎ | changedFilesWithAncestor
+    ✖︎ | changedSince
+    ✔︎ | collectCoverage
+    ✔︎ | collectCoverageFrom
+    ✔︎ | collectCoverageOnlyFrom
+    ✔︎ | coverageDirectory
+    ✔︎ | coverageReporters
+    ✖︎ | coverageThreshold
+    ✖︎ | detectLeaks
+    ✖︎ | detectOpenHandles
+    ✖︎ | enabledTestsMap
+    ✖︎ | errorOnDeprecated
+    ✖︎ | expand
+    ✖︎ | filter
+    ✖︎ | findRelatedTests
+    ✖︎ | forceExit
+    ✖︎ | globalSetup
+    ✖︎ | globalTeardown
+    ✖︎ | json
+    ✖︎ | lastCommit
+    ✖︎ | listTests
+    ✖︎ | logHeapUsage
+    ✖︎ | maxWorkers
+    ✖︎ | nonFlagArgs
+    ✖︎ | noSCM
+    ✖︎ | noStackTrace
+    ✔︎ | notify
+    ✔︎ | notifyMode
+    ✖︎ | onlyChanged
+    ✔︎ | onlyFailures
+    ✖︎ | outputFile
+    ✖︎ | passWithNoTests
+    ✖︎ | projects
+    ✖︎ | replname
+    ✔︎ | reporters
+    ✖︎ | rootDir
+    ✖︎ | runTestsByPath
+    ✖︎ | silent
+    ✖︎ | skipFilter
+    ✖︎ | testFailureExitCode
+    ✔︎ | testNamePattern
+    ✔︎ | testPathPattern
+    ✖︎ | testResultsProcessor
+    ✔︎ | updateSnapshot
+    ✖︎ | useStderr
+    ✔︎ | verbose
+    ✖︎ | watch
+    ✖︎ | watchAll
+    ✖︎ | watchman
+    ✖︎ | watchPlugins
+  `(
+    'allows WatchPlugins to modify only white-listed global config keys',
+    async ({ok, option}) => {
+      const pluginPath = `${__dirname}/__fixtures__/plugin_path_config_updater`;
+      const config = Object.assign({}, globalConfig, {
+        rootDir: __dirname,
+        watchPlugins: [{config: {}, path: pluginPath}],
+      });
+
+      jest.doMock(
+        pluginPath,
+        () =>
+          class WatchPlugin {
+            getUsageInfo() {
+              return {key: 'x', prompt: 'test option white-listing'};
+            }
+
+            run(globalConfig, updateConfigAndRun) {
+              updateConfigAndRun({[option]: '__JUST_TRYING__'});
+              return Promise.resolve();
+            }
+          },
+        {virtual: true},
+      );
+
+      watch(config, contexts, pipe, hasteMapInstances, stdin);
+      await nextTick();
+
+      stdin.emit('x');
+      await nextTick();
+
+      const lastCall = updateGlobalConfig.mock.calls.slice(-1)[0];
+      let expector = expect(lastCall[0]);
+      if (!ok) {
+        expector = expector.not;
+      }
+      expector.toMatchObject({
+        [option]: '__JUST_TRYING__',
+      });
+    },
+  );
 
   it('triggers enter on a WatchPlugin when its key is pressed', async () => {
     const run = jest.fn(() => Promise.resolve());
@@ -414,7 +558,7 @@ describe('Watch mode flows', () => {
     watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [pluginPath],
+        watchPlugins: [{config: {}, path: pluginPath}],
       }),
       contexts,
       pipe,
@@ -474,7 +618,10 @@ describe('Watch mode flows', () => {
     watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
-        watchPlugins: [pluginPath, pluginPath2],
+        watchPlugins: [
+          {config: {}, path: pluginPath},
+          {config: {}, path: pluginPath2},
+        ],
       }),
       contexts,
       pipe,
@@ -541,7 +688,6 @@ describe('Watch mode flows', () => {
 
     expect(runJestMock.mock.calls[0][0].globalConfig).toMatchObject({
       testNamePattern: 'test',
-      testPathPattern: '',
       watch: true,
       watchAll: false,
     });
@@ -559,7 +705,6 @@ describe('Watch mode flows', () => {
     await nextTick();
 
     expect(runJestMock.mock.calls[0][0].globalConfig).toMatchObject({
-      testNamePattern: '',
       testPathPattern: 'file',
       watch: true,
       watchAll: false,