+++ packages/jest-jasmine2/src/setup_jest_globals.js
@@ -98,7 +98,12 @@ export default ({
     });
   patchJasmine();
   const {expand, updateSnapshot} = globalConfig;
-  const snapshotState = new SnapshotState(testPath, {expand, updateSnapshot});
+  const snapshotState = new SnapshotState(testPath, {
+    expand,
+    getPrettier: () =>
+      config.prettierPath ? localRequire(config.prettierPath) : null,
+    updateSnapshot,
+  });
   setState({snapshotState, testPath});
   // Return it back to the outer scope (test runner outside the VM).
   return snapshotState;