+++ website/pages/en/index.js
@@ -14,8 +14,11 @@ const GridBlock = CompLibrary.GridBlock;
 
 const translate = require('../../server/translate.js').translate;
 const translation = require('../../server/translation.js');
+const backers = require(process.cwd() + '/backers.json');
 
 const siteConfig = require(process.cwd() + '/siteConfig.js');
+const idx = (target, path) =>
+  path.reduce((obj, key) => (obj && obj[key] ? obj[key] : null), target);
 
 class Button extends React.Component {
   render() {
@@ -33,6 +36,89 @@ Button.defaultProps = {
   target: '_self',
 };
 
+class Contributors extends React.Component {
+  render() {
+    return (
+      <div>
+        <h2>
+          <translate>Sponsors</translate>
+        </h2>
+        <p>
+          <translate>
+            Sponsors are those who contribute $100 or more per month to Jest
+          </translate>
+        </p>
+        <div>
+          {backers.filter(b => b.tier === 'sponsor').map(b => (
+            <a
+              key={b.id}
+              className="sponsor-item"
+              title={`$${b.totalDonations / 100} by ${b.name || b.slug}`}
+              target="_blank"
+              href={b.website || `https://opencollective.com/${b.slug}`}
+            >
+              {
+                <img
+                  className="sponsor-avatar"
+                  src={b.avatar}
+                  alt={
+                    b.name || b.slug ? `${b.name || b.slug}'s avatar` : 'avatar'
+                  }
+                />
+              }
+            </a>
+          ))}
+        </div>
+        <div className="support">
+          <a
+            className="support-button"
+            href="https://opencollective.com/jest#support"
+          >
+            <translate>Become a sponsor</translate>
+          </a>
+        </div>
+        <h2>
+          <translate>Backers</translate>
+        </h2>
+        <p>
+          <translate>
+            Backers are those who contribute $2 or more per month to Jest
+          </translate>
+        </p>
+        <div>
+          {backers.filter(b => b.tier === 'backer').map(b => (
+            <a
+              key={b.id}
+              className="backer-item"
+              title={`$${b.totalDonations / 100} by ${b.name || b.slug}`}
+              target="_blank"
+              href={b.website || `https://opencollective.com/${b.slug}`}
+            >
+              {
+                <img
+                  className="backer-avatar"
+                  src={b.avatar}
+                  alt={
+                    b.name || b.slug ? `${b.name || b.slug}'s avatar` : 'avatar'
+                  }
+                />
+              }
+            </a>
+          ))}
+          <div>
+            <a
+              className="support-button"
+              href="https://opencollective.com/jest#support"
+            >
+              <translate>Become a backer</translate>
+            </a>
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
+
 class HomeSplash extends React.Component {
   render() {
     return (
@@ -46,10 +132,11 @@ class HomeSplash extends React.Component {
               <h2 className="projectTitle">
                 {siteConfig.title}
                 <small>
-                  {
-                    translation[this.props.language]['localized-strings']
-                      .tagline
-                  }
+                  {idx(translation, [
+                    this.props.language,
+                    'localized-strings',
+                    'tagline',
+                  ]) || siteConfig.tagline}
                 </small>
               </h2>
               <div className="section promoSection">
@@ -128,7 +215,7 @@ class Index extends React.Component {
                       Works out of the box for any React project.
                     </translate>
                   ),
-                  image: '/jest/img/content/female-technologist.png',
+                  image: '/img/content/female-technologist.png',
                   imageAlign: 'top',
                   title: <translate>Developer Ready</translate>,
                 },
@@ -139,7 +226,7 @@ class Index extends React.Component {
                       to changed files and is optimized to give signal quickly.
                     </translate>
                   ),
-                  image: '/jest/img/content/runner.png',
+                  image: '/img/content/runner.png',
                   imageAlign: 'top',
                   title: <translate>Instant Feedback</translate>,
                 },
@@ -151,7 +238,7 @@ class Index extends React.Component {
                       changes over time.
                     </translate>
                   ),
-                  image: '/jest/img/content/camera-with-flash.png',
+                  image: '/img/content/camera-with-flash.png',
                   imageAlign: 'top',
                   title: <translate>Snapshot Testing</translate>,
                 },
@@ -192,7 +279,7 @@ class Index extends React.Component {
                       tests conflict with each other.
                     </translate>
                   ),
-                  image: '/jest/img/content/feature-fast.png',
+                  image: '/img/content/feature-fast.png',
                   imageAlign: 'right',
                   title: <translate>Fast and sandboxed</translate>,
                 },
@@ -206,13 +293,13 @@ class Index extends React.Component {
                   content: (
                     <translate>
                       Easily create code coverage reports using
-                      [`--coverage`](https://facebook.github.io/jest/docs/en/cli.html#coverage).
+                      [`--coverage`](https://jestjs.io/docs/en/cli.html#coverage).
                       No additional setup or libraries needed! Jest can collect
                       code coverage information from entire projects, including
                       untested files.
                     </translate>
                   ),
-                  image: '/jest/img/content/feature-coverage.png',
+                  image: '/img/content/feature-coverage.png',
                   imageAlign: 'left',
                   title: <translate>Built-in code coverage reports</translate>,
                 },
@@ -235,7 +322,7 @@ class Index extends React.Component {
                       prefer, Jest will find and run your tests.
                     </translate>
                   ),
-                  image: '/jest/img/content/feature-config-react.png',
+                  image: '/img/content/feature-config-react.png',
                   imageAlign: 'right',
                   title: <translate>Zero configuration</translate>,
                 },
@@ -274,13 +361,12 @@ class Index extends React.Component {
                 {
                   content: (
                     <translate>
-                      Powerful [mocking
-                      library](/jest/docs/en/mock-functions.html) for functions
-                      and modules. Mock React Native components using
-                      `jest-react-native`.
+                      Powerful [mocking library](/docs/en/mock-functions.html)
+                      for functions and modules. Mock React Native components
+                      using `jest-react-native`.
                     </translate>
                   ),
-                  image: '/jest/img/content/feature-mocking.png',
+                  image: '/img/content/feature-mocking.png',
                   imageAlign: 'left',
                   title: <translate>Powerful mocking library</translate>,
                 },
@@ -300,7 +386,7 @@ class Index extends React.Component {
                       [ts-jest](https://github.com/kulshekhar/ts-jest).
                     </translate>
                   ),
-                  image: '/jest/img/content/feature-typescript.png',
+                  image: '/img/content/feature-typescript.png',
                   imageAlign: 'right',
                   title: <translate>Works with TypeScript</translate>,
                 },
@@ -367,6 +453,7 @@ class Index extends React.Component {
               </translate>
             </p>
             <div className="logos">{showcase}</div>
+            <Contributors />
           </div>
         </div>
       </div>